requirements.txt
absl-py==2.0.0
albumentations==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
chumpy @ file:///mnt/newdisk/python/projects/december/3d-model/chumpy
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
Cython==3.0.6
daphne==4.0.0
Django==4.2.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.3.1
djangorestframework==3.14.0
efficientnet-pytorch==0.7.1
face-alignment==1.4.1
filelock==3.13.1
filterpy==1.4.5
flake8==6.1.0
fonttools==4.45.1
fsspec==2023.10.0
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.7.1
google-auth==2.25.1
google-auth-oauthlib==1.1.0
grpcio==1.60.0
gunicorn==21.2.0
h5py==3.10.0
huggingface-hub==0.19.4
hyperlink==21.0.0
idna==3.6
imageio==2.33.0
importlib-metadata==7.0.0
incremental==22.10.0
iopath==0.1.10
isort==4.3.21
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.4
kornia==0.7.0
lazy_loader==0.3
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mccabe==0.7.0
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.1.0
portalocker==2.8.2
pretrainedmodels==0.7.4
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
segmentation-models-pytorch==0.3.3
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tb-nightly==2.16.0a20231206
tensorboard-data-server==0.7.2
termcolor==2.4.0
tf-keras-nightly==2.16.0.dev2023120710
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.2
tomli==2.0.1
torch==2.1.2
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.220
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
yacs==0.1.8
yapf==0.40.2
zipp==3.17.0
zope.interface==6.1
